chore: rename analytics
