+++ internal/storage/analytics/file.go
@@ -1,4 +1,4 @@
-package sql
+package analytics
 
 import (
 	"io/fs"
