+++ internal/storage/analytics/migrator_test.go
@@ -1,4 +1,4 @@
-package sql
+package analytics
 
 import (
 	"os"
